requirements.txt
Django==5.0.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
